requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
loguru==0.7.2
apscheduler==3.10.4
aiohttp==3.9.3
